refactor: QuizType 모듈 이동
